requirements.txt

azure-storage-blob>=12.0.0
requests>=2.0.0
tenacity>=8.0.0
python-dateutil>=2.8.0
